LayoutTests/inspector/profiler/memory-instrumentation-cached-images.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script>

function displayAndRunTest()
{
    if (window.testRunner)
        testRunner.display();
    runTest();
}


function test()
{
    function memoryBlockSize(path, root)
    {
        var pathPos = 0;
        var children = [root];

        while (true) {
            var name = path[pathPos++];
            var child = null;
            for (var i = 0; i < children.length; i++) {
                if (children[i].name === name) {
                    if (pathPos === path.length)
                        return children[i].size;
                    else {
                        child = children[i];
                        break;
                    }
                }
            }
            if (child) {
                children = child.children;
                if (!children) {
                    InspectorTest.addResult(name + " has no children");
                    return -1;
                }
            } else {
                InspectorTest.addResult(name + " not found");
                return -1;
            }
        }
        return -1;
    }

    function didReceiveMemorySnapshot(error, memoryBlock)
    {
        InspectorTest.addResult("didReceiveMemorySnapshot");
        var imagesSize = memoryBlockSize(["ProcessPrivateMemory", "MemoryCache", "CachedImages"], memoryBlock);
        var jpegDecodedSizeExpected = 20000000;
        if (imagesSize > jpegDecodedSizeExpected)
            InspectorTest.addResult("PASS: CachedImages size is OK.");
        else
            InspectorTest.addResult("FAIL: CachedImages size is too small.");
        InspectorTest.completeTest();
    }

    MemoryAgent.getProcessMemoryDistribution(didReceiveMemorySnapshot.bind(this));
}

</script>
</head>

<body>
<p>
This test checks that CachedImages size reported by the memory agent includes images decoded size.
<a href="https://bugs.webkit.org/show_bug.cgi?id=93366">Bug 93366</a>
</p>
<img src="resources/image.jpeg" width="100" height="60" onload="displayAndRunTest()"/>
</body>
</html>
